+++ features/groups/src/main/kotlin/es/pedrazamiguez/splittrip/features/group/presentation/component/CameraPreview.kt
@@ -1,0 +1,111 @@
+package es.pedrazamiguez.splittrip.features.group.presentation.component
+
+import androidx.annotation.OptIn
+import androidx.camera.core.CameraSelector
+import androidx.camera.core.ExperimentalGetImage
+import androidx.camera.core.ImageAnalysis
+import androidx.camera.core.ImageProxy
+import androidx.camera.core.Preview
+import androidx.camera.lifecycle.ProcessCameraProvider
+import androidx.camera.view.PreviewView
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalLifecycleOwner
+import androidx.compose.ui.viewinterop.AndroidView
+import androidx.core.content.ContextCompat
+import com.google.mlkit.vision.barcode.BarcodeScanning
+import com.google.mlkit.vision.common.InputImage
+import es.pedrazamiguez.splittrip.core.common.util.QrCodePayloadParser
+import java.util.concurrent.Executors
+
+@Composable
+internal fun CameraPreview(
+    onScanned: (QrCodePayloadParser.SharePayload) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val lifecycleOwner = LocalLifecycleOwner.current
+    val cameraExecutor = remember { Executors.newSingleThreadExecutor() }
+
+    DisposableEffect(Unit) {
+        onDispose {
+            cameraExecutor.shutdown()
+        }
+    }
+
+    AndroidView(
+        factory = { ctx ->
+            val previewView = PreviewView(ctx)
+            val cameraProviderFuture = ProcessCameraProvider.getInstance(ctx)
+
+            cameraProviderFuture.addListener({
+                val cameraProvider = cameraProviderFuture.get()
+                val preview = Preview.Builder().build().apply {
+                    surfaceProvider = previewView.surfaceProvider
+                }
+
+                val imageAnalysis = ImageAnalysis.Builder()
+                    .setBackpressureStrategy(ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST)
+                    .build()
+
+                imageAnalysis.setAnalyzer(
+                    cameraExecutor,
+                    QrCodeAnalyzer(onScanned = onScanned)
+                )
+
+                val cameraSelector = CameraSelector.DEFAULT_BACK_CAMERA
+
+                try {
+                    cameraProvider.unbindAll()
+                    cameraProvider.bindToLifecycle(
+                        lifecycleOwner,
+                        cameraSelector,
+                        preview,
+                        imageAnalysis
+                    )
+                } catch (ignored: Exception) {
+                    // Handled gracefully, camera binding failed
+                }
+            }, ContextCompat.getMainExecutor(ctx))
+
+            previewView
+        },
+        modifier = modifier.fillMaxSize()
+    )
+}
+
+private class QrCodeAnalyzer(
+    private val onScanned: (QrCodePayloadParser.SharePayload) -> Unit
+) : ImageAnalysis.Analyzer {
+
+    private val barcodeScanner = BarcodeScanning.getClient()
+    private var isScanningActive = true
+
+    @OptIn(ExperimentalGetImage::class)
+    override fun analyze(imageProxy: ImageProxy) {
+        val mediaImage = imageProxy.image
+        if (mediaImage != null && isScanningActive) {
+            val image = InputImage.fromMediaImage(mediaImage, imageProxy.imageInfo.rotationDegrees)
+            barcodeScanner.process(image)
+                .addOnSuccessListener { barcodes ->
+                    val payload = barcodes.firstNotNullOfOrNull { barcode ->
+                        barcode.rawValue?.let { QrCodePayloadParser.parseSharePayload(it) }
+                    }
+                    if (payload != null) {
+                        isScanningActive = false
+                        onScanned(payload)
+                    }
+                }
+                .addOnFailureListener {
+                    // Ignore scanner failures
+                }
+                .addOnCompleteListener {
+                    imageProxy.close()
+                }
+        } else {
+            imageProxy.close()
+        }
+    }
+}
