+++ features/groups/src/main/kotlin/es/pedrazamiguez/splittrip/features/group/presentation/component/QrScannerDialog.kt
@@ -1,19 +1,10 @@
 package es.pedrazamiguez.splittrip.features.group.presentation.component
 
-import androidx.annotation.OptIn
-import androidx.camera.core.CameraSelector
-import androidx.camera.core.ExperimentalGetImage
-import androidx.camera.core.ImageAnalysis
-import androidx.camera.core.ImageProxy
-import androidx.camera.core.Preview
-import androidx.camera.lifecycle.ProcessCameraProvider
-import androidx.camera.view.PreviewView
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -25,25 +16,18 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
-import androidx.core.content.ContextCompat
-import com.google.mlkit.vision.barcode.BarcodeScanning
-import com.google.mlkit.vision.common.InputImage
 import es.pedrazamiguez.splittrip.core.common.util.QrCodePayloadParser
+import es.pedrazamiguez.splittrip.core.designsystem.foundation.spacing
 import es.pedrazamiguez.splittrip.features.group.R
-import java.util.concurrent.Executors
 
 @Composable
 internal fun QrScannerDialog(
@@ -58,14 +42,14 @@ internal fun QrScannerDialog(
         Card(
             modifier = modifier
                 .fillMaxWidth()
-                .padding(24.dp),
+                .padding(MaterialTheme.spacing.ExtraLarge),
             shape = RoundedCornerShape(16.dp),
             colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)
         ) {
             Column(
                 modifier = Modifier
                     .fillMaxWidth()
-                    .padding(16.dp),
+                    .padding(MaterialTheme.spacing.Default),
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
                 Text(
@@ -73,14 +57,14 @@ internal fun QrScannerDialog(
                     style = MaterialTheme.typography.titleLarge,
                     color = MaterialTheme.colorScheme.onSurface
                 )
-                Spacer(modifier = Modifier.height(8.dp))
+                Spacer(modifier = Modifier.height(MaterialTheme.spacing.Small))
                 Text(
                     text = stringResource(R.string.scanner_helper_text),
                     style = MaterialTheme.typography.bodyMedium,
                     color = MaterialTheme.colorScheme.onSurfaceVariant,
                     textAlign = TextAlign.Center
                 )
-                Spacer(modifier = Modifier.height(16.dp))
+                Spacer(modifier = Modifier.height(MaterialTheme.spacing.Default))
                 Box(
                     modifier = Modifier
                         .size(250.dp)
@@ -95,100 +79,11 @@ internal fun QrScannerDialog(
                             .border(2.dp, MaterialTheme.colorScheme.primary, RoundedCornerShape(8.dp))
                     )
                 }
-                Spacer(modifier = Modifier.height(16.dp))
+                Spacer(modifier = Modifier.height(MaterialTheme.spacing.Default))
                 TextButton(onClick = onDismissRequest) {
                     Text(text = stringResource(R.string.scanner_close))
                 }
             }
-        }
-    }
-}
-
-@Composable
-private fun CameraPreview(
-    onScanned: (QrCodePayloadParser.SharePayload) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val lifecycleOwner = LocalLifecycleOwner.current
-    val cameraExecutor = remember { Executors.newSingleThreadExecutor() }
-
-    DisposableEffect(Unit) {
-        onDispose {
-            cameraExecutor.shutdown()
-        }
-    }
-
-    AndroidView(
-        factory = { ctx ->
-            val previewView = PreviewView(ctx)
-            val cameraProviderFuture = ProcessCameraProvider.getInstance(ctx)
-
-            cameraProviderFuture.addListener({
-                val cameraProvider = cameraProviderFuture.get()
-                val preview = Preview.Builder().build().apply {
-                    surfaceProvider = previewView.surfaceProvider
-                }
-
-                val imageAnalysis = ImageAnalysis.Builder()
-                    .setBackpressureStrategy(ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST)
-                    .build()
-
-                imageAnalysis.setAnalyzer(
-                    cameraExecutor,
-                    QrCodeAnalyzer(onScanned = onScanned)
-                )
-
-                val cameraSelector = CameraSelector.DEFAULT_BACK_CAMERA
-
-                try {
-                    cameraProvider.unbindAll()
-                    cameraProvider.bindToLifecycle(
-                        lifecycleOwner,
-                        cameraSelector,
-                        preview,
-                        imageAnalysis
-                    )
-                } catch (ignored: Exception) {
-                    // Handled gracefully, camera binding failed
-                }
-            }, ContextCompat.getMainExecutor(ctx))
-
-            previewView
-        },
-        modifier = modifier.fillMaxSize()
-    )
-}
-
-private class QrCodeAnalyzer(
-    private val onScanned: (QrCodePayloadParser.SharePayload) -> Unit
-) : ImageAnalysis.Analyzer {
-
-    private val barcodeScanner = BarcodeScanning.getClient()
-    private var isScanningActive = true
-
-    @OptIn(ExperimentalGetImage::class)
-    override fun analyze(imageProxy: ImageProxy) {
-        val mediaImage = imageProxy.image
-        if (mediaImage != null && isScanningActive) {
-            val image = InputImage.fromMediaImage(mediaImage, imageProxy.imageInfo.rotationDegrees)
-            barcodeScanner.process(image)
-                .addOnSuccessListener { barcodes ->
-                    val payload = barcodes.firstNotNullOfOrNull { barcode ->
-                        barcode.rawValue?.let { QrCodePayloadParser.parseSharePayload(it) }
-                    }
-                    if (payload != null) {
-                        isScanningActive = false
-                        onScanned(payload)
-                    }
-                }
-                .addOnFailureListener {
-                    // Ignore scanner failures
-                }
-                .addOnCompleteListener {
-                    imageProxy.close()
-                }
-        } else {
-            imageProxy.close()
         }
     }
 }
